refactor(pack): refactor xdsl dhcp
