+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/router/dhcp/dhcp-modal.component.js
@@ -1,0 +1,15 @@
+export default /* @ngInject */ class telecomXdslModemDhcpModal {
+  /* @ngInject */
+  constructor($uibModalInstance, messageIps) {
+    this.$uibModalInstance = $uibModalInstance;
+    this.messageIps = messageIps;
+  }
+
+  cancel(message) {
+    return this.$uibModalInstance.dismiss(message);
+  }
+
+  close() {
+    return this.$uibModalInstance.close(true);
+  }
+}
