--- packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/router/dhcp/pack-xdsl-modem-dhcp-modal.component.js
@@ -1,18 +1,0 @@
-angular.module('managerApp').controller(
-  'telecomXdslModemDhcpModal',
-  class telecomXdslModemDhcpModal {
-    /* @ngInject */
-    constructor($uibModalInstance, messageIps) {
-      this.$uibModalInstance = $uibModalInstance;
-      this.messageIps = messageIps;
-    }
-
-    cancel(message) {
-      return this.$uibModalInstance.dismiss(message);
-    }
-
-    close() {
-      return this.$uibModalInstance.close(true);
-    }
-  },
-);
